requirements.txt
pandas
pillow
yagmail
openpyxl
